+++ 1.21.8/Common/src/main/java/fuzs/puzzleslib/api/data/v2/core/DataProviderContext.java
@@ -38,7 +38,7 @@ public class DataProviderContext {
      * @param packOutput the pack output
      * @param registries the registry lookup provider
      */
-    protected DataProviderContext(String modId, PackOutput packOutput, CompletableFuture<HolderLookup.Provider> registries) {
+    public DataProviderContext(String modId, PackOutput packOutput, CompletableFuture<HolderLookup.Provider> registries) {
         this(modId, packOutput, () -> registries);
     }