+++ 1.21.8/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/api/event/v1/core/NeoForgeEventInvokerRegistry.java
@@ -8,7 +8,6 @@ import net.neoforged.bus.api.Event;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.function.BiConsumer;
-import java.util.function.Function;
 import java.util.function.UnaryOperator;
 
 /**
@@ -24,64 +23,70 @@ public interface NeoForgeEventInvokerRegistry extends EventInvokerRegistry {
     /**
      * Registers an event.
      *
-     * @param clazz     common event functional interface class
-     * @param event     NeoForge event implementation
-     * @param converter passes parameters from the NeoForge event to the common event instance
-     * @param <T>       common event type
-     * @param <E>       NeoForge event type
+     * @param clazz      common event functional interface class
+     * @param eventClazz NeoForge event implementation
+     * @param converter  passes parameters from the NeoForge event to the common event instance
+     * @param <T>        common event type
+     * @param <E>        NeoForge event type
      */
-    default <T, E extends Event> void register(Class<T> clazz, Class<E> event, BiConsumer<T, E> converter) {
-        this.register(clazz, event, (T callback, E evt, Object context) -> converter.accept(callback, evt), false);
+    default <T, E extends Event> void register(Class<T> clazz, Class<E> eventClazz, BiConsumer<T, E> converter) {
+        this.register(clazz,
+                eventClazz,
+                (T callback, E event, Object context) -> converter.accept(callback, event),
+                false);
     }
 
     /**
      * Registers an event.
      *
      * @param clazz        common event functional interface class
-     * @param event        NeoForge event implementation
+     * @param eventClazz   NeoForge event implementation
      * @param converter    passes parameters from the NeoForge event to the common event instance
      * @param joinInvokers join this new event invoker with a possibly already existing one, otherwise an exception will
      *                     be thrown when registering duplicates
      * @param <T>          common event type
      * @param <E>          NeoForge event type
      */
-    default <T, E extends Event> void register(Class<T> clazz, Class<E> event, BiConsumer<T, E> converter, boolean joinInvokers) {
-        this.register(clazz, event, (T callback, E evt, @Nullable Object context) -> converter.accept(callback, evt), joinInvokers);
+    default <T, E extends Event> void register(Class<T> clazz, Class<E> eventClazz, BiConsumer<T, E> converter, boolean joinInvokers) {
+        this.register(clazz,
+                eventClazz,
+                (T callback, E event, @Nullable Object context) -> converter.accept(callback, event),
+                joinInvokers);
     }
 
     /**
      * Registers an event.
      *
-     * @param clazz     common event functional interface class
-     * @param event     NeoForge event implementation
-     * @param converter passes parameters from the NeoForge event to the common event instance, including a context
-     * @param <T>       common event type
-     * @param <E>       NeoForge event type
+     * @param clazz      common event functional interface class
+     * @param eventClazz NeoForge event implementation
+     * @param converter  passes parameters from the NeoForge event to the common event instance, including a context
+     * @param <T>        common event type
+     * @param <E>        NeoForge event type
      */
-    default <T, E extends Event> void register(Class<T> clazz, Class<E> event, NeoForgeEventContextConsumer<T, E> converter) {
-        this.register(clazz, event, converter, false);
+    default <T, E extends Event> void register(Class<T> clazz, Class<E> eventClazz, NeoForgeEventContextConsumer<T, E> converter) {
+        this.register(clazz, eventClazz, converter, false);
     }
 
     /**
      * Registers an event.
      *
      * @param clazz        common event functional interface class
-     * @param event        NeoForge event implementation
+     * @param eventClazz   NeoForge event implementation
      * @param converter    passes parameters from the NeoForge event to the common event instance, including a context
      * @param joinInvokers join this new event invoker with a possibly already existing one, otherwise an exception will
      *                     be thrown when registering duplicates
      * @param <T>          common event type
      * @param <E>          NeoForge event type
      */
-    default <T, E extends Event> void register(Class<T> clazz, Class<E> event, NeoForgeEventContextConsumer<T, E> converter, boolean joinInvokers) {
-        this.register(clazz, event, converter, UnaryOperator.identity(), joinInvokers);
+    default <T, E extends Event> void register(Class<T> clazz, Class<E> eventClazz, NeoForgeEventContextConsumer<T, E> converter, boolean joinInvokers) {
+        this.register(clazz, eventClazz, converter, UnaryOperator.identity(), joinInvokers);
     }
 
     /**
      * Registers an event.
      *
      * @param clazz               common event functional interface class
-     * @param event               NeoForge event implementation
+     * @param eventClazz          NeoForge event implementation
      * @param converter           passes parameters from the NeoForge event to the common event instance, including a
      *                            context
      * @param eventPhaseConverter an operator for adjusting the provided event phase, intended to impose an ordering on
@@ -91,7 +96,7 @@ public interface NeoForgeEventInvokerRegistry extends EventInvokerRegistry {
      * @param <T>                 common event type
      * @param <E>                 NeoForge event type
      */
-    <T, E extends Event> void register(Class<T> clazz, Class<E> event, NeoForgeEventContextConsumer<T, E> converter, UnaryOperator<EventPhase> eventPhaseConverter, boolean joinInvokers);
+    <T, E extends Event> void register(Class<T> clazz, Class<E> eventClazz, NeoForgeEventContextConsumer<T, E> converter, UnaryOperator<EventPhase> eventPhaseConverter, boolean joinInvokers);
 
     /**
      * A helper context for dealing with context based {@link EventInvoker} implementations.